requirements.txt
PySide6>=6.6.0
requests>=2.31.0
PyYAML>=6.0.1
Pillow>=10.0.0
